save towns only for authors and admins

+++ ckanext/ctdata_theme/ctdata/location/controllers.py
@@ -31,6 +31,7 @@ class LocationsController(base.BaseController):
         self.community_profile_service = CommunityProfileService(self.session)
         self.user_service     = UserService(self.session)
         self.location_service = LocationService(self.session)
+    #end
 
     def show(self, location_name):
         location = self.location_service.get_location(location_name)
@@ -45,16 +46,19 @@ class LocationsController(base.BaseController):
             self.session.add(location_profile)
 
         return base.render('location/show.html', extra_vars={'location': location, 'towns': towns, 'towns_names': towns_names, 'default_profile_id': default_profile.id})
+    #end
 
     def data_by_location(self):
         locations = self.location_service.get_all_locations()
 
         return base.render('location/data_by_location.html', extra_vars={'locations': locations})
+    #end
 
     def manage_locations(self):
         locations = self.location_service.get_all_locations()
 
         return base.render('location/manage_locations.html', extra_vars={'locations': locations})
+    #end
 
     def create_location(self):
         if http_request.method == 'POST':
@@ -66,11 +70,13 @@ class LocationsController(base.BaseController):
 
         http_response.headers['Content-type'] = 'application/json'
         return json.dumps({'location_name': location.name, 'location_fips': location.fips})
+    #end
 
     def new_profile(self, location_name):
         location = self.location_service.get_location(location_name)
 
         return base.render('location/new_profile.html', extra_vars={'location': location})
+    #end
 
     def load_indicator(self, location_name):
         location = self.location_service.get_location(location_name)
@@ -123,7 +129,7 @@ class LocationsController(base.BaseController):
         return json.dumps({'success': False, 'error': str('Indicator cannot be saved')})
 
         return base.render('location/new_profile.html', extra_vars={'location': location})
-
+    #end
 
     def create_location_profile(self, location_name):
         http_response.headers['Content-type'] = 'application/json'
@@ -145,6 +151,7 @@ class LocationsController(base.BaseController):
             return json.dumps({'success': True,  'redirect_link': redirect_link})
 
         return json.dumps({'success': False, 'error': str('Profile cannot be saved')})
+    #end
 
     def community_profile(self, profile_id):
         default_profile = self.location_service.get_profile(profile_id)
@@ -152,36 +159,46 @@ class LocationsController(base.BaseController):
         towns_names = ','.join( l for l in map(lambda t: t.name, default_profile.locations))
 
         return base.render('location/show.html', extra_vars={'location': default_profile.locations[0], 'towns': towns, 'towns_names': towns_names, 'default_profile_id': default_profile.id})
+    #end
 
     def load_profile_indicators(self, profile_id):
         http_response.headers['Content-type'] = 'application/json'
 
+        session_id     = session.id
+        user_name      = http_request.environ.get("REMOTE_USER") or "guest_" + session_id
         json_body      = json.loads(http_request.body, encoding=http_request.charset)
         location_names = json_body.get('locations')
         location_names = location_names.split(',')
         profile        = self.location_service.get_profile(profile_id)
+        user           = self.user_service.get_or_create_user_with_session_id(user_name,session_id) if user_name else None
         locations      = []
 
-        ######### save new towns
-        for profile_location in profile.locations:
-            if profile_location.name not in location_names:
-                self.location_service.remove_location_profile(profile_location.id, profile.id)
-                profile.locations.remove(profile_location)
+        if (not profile.user and user.is_admin) or profile.user == user:
+            ######### save new towns
+            for profile_location in profile.locations:
+                if profile_location.name not in location_names:
+                    self.location_service.remove_location_profile(profile_location.id, profile.id)
+                    profile.locations.remove(profile_location)
 
-        for location_name in location_names:
-            location = self.location_service.get_location(location_name)
-            locations.append( location )
-            if location not in profile.locations:
-                profile.locations.append(location)
-                location_profile = LocationProfile(location.id, profile.id)
-                self.session.add(location_profile)
+            for location_name in location_names:
+                location = self.location_service.get_location(location_name)
+                locations.append( location )
+                if location not in profile.locations:
+                    profile.locations.append(location)
+                    location_profile = LocationProfile(location.id, profile.id)
+                    self.session.add(location_profile)
 
-        self.session.commit()
+            self.session.commit()
+            locations_names = map(lambda t: t.name, profile.locations)
+        else:
+            locations_names = location_names
+
 
         ######### load indicators data
+
         ind_data = []
         for indicator in profile.indicators:
-            values = self.location_service.load_indicator_value_for_location(indicator.filters, indicator.dataset_id, map(lambda t: t.name, profile.locations))
+            values = self.location_service.load_indicator_value_for_location(indicator.filters, indicator.dataset_id, locations_names)
 
             data  = {}
             data  = {
@@ -198,7 +215,8 @@ class LocationsController(base.BaseController):
 
             ind_data.append(data)
 
-        return json.dumps({'success': True, 'ind_data': ind_data, 'towns':  map(lambda t: t.name, profile.locations)})
+        return json.dumps({'success': True, 'ind_data': ind_data, 'towns':  locations_names})
+    #end
 
     def save_local_default(self, profile_id):
         http_response.headers['Content-type'] = 'application/json'
@@ -218,4 +236,4 @@ class LocationsController(base.BaseController):
 
         #remove deleted indicators
         return json.dumps({'success': True })
-
+    #end
